refactor: move hull to points.hull
